Solution to Check Multiples

+++ Check Multiples@@6788dd63490637cfc44e1b9a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    int a,b;
-    scanf("%d %d",&a,&b);
-    if (b/a==0){
+    float a,b;
+    scanf("%f %f",&a,&b);
+    if (b/a==0.00){
         printf("Yes");
     }
     else{
